Move Docker utilities out of `splitgraph.commandline` to avoid circular imports (since plugins are loaded in the commandline module).
