+++ splitgraph/commandline/engine.py
@@ -1,11 +1,8 @@
 import logging
 import os
 import platform
-import time
-from io import BytesIO
 from pathlib import Path, PureWindowsPath
-from tarfile import TarFile, TarInfo
-from typing import Dict, TYPE_CHECKING, Optional
+from typing import Dict, TYPE_CHECKING
 from urllib.parse import urlparse
 
 import click
@@ -14,60 +11,13 @@ from tqdm import tqdm
 from splitgraph.__version__ import __version__
 from splitgraph.config import CONFIG, SG_CMD_ASCII
 from splitgraph.exceptions import DockerUnavailableError, EngineSetupError
+from splitgraph.utils.docker import get_docker_client, copy_to_container
 
 if TYPE_CHECKING:
-    from docker.models.containers import Container
+    pass
 
 
 DEFAULT_ENGINE = "default"
-
-
-def get_docker_client():
-    """Wrapper around client.from_env() that also pings the daemon
-    to make sure it can connect and if not, raises an error."""
-    import docker
-
-    try:
-        client = docker.from_env()
-        client.ping()
-        return client
-    except Exception as e:
-        raise DockerUnavailableError("Could not connect to the Docker daemon") from e
-
-
-def copy_to_container(
-    container: "Container",
-    source_path: Optional[str],
-    target_path: str,
-    data: Optional[bytes] = None,
-) -> None:
-    """
-    Copy a file into a Docker container
-
-    :param container: Container object
-    :param source_path: Source file path
-    :param target_path: Target file path (in the container)
-    :return:
-    """
-
-    if data is None:
-        if not source_path:
-            raise ValueError("One of source_path or data must be specified!")
-        # https://github.com/docker/docker-py/issues/1771
-        with open(source_path, "rb") as f:
-            data = f.read()
-
-    tarinfo = TarInfo(name=os.path.basename(target_path))
-    tarinfo.size = len(data)
-    tarinfo.mtime = int(time.time())
-
-    stream = BytesIO()
-    tar = TarFile(fileobj=stream, mode="w")
-    tar.addfile(tarinfo, BytesIO(data))
-    tar.close()
-
-    stream.seek(0)
-    container.put_archive(path=os.path.dirname(target_path), data=stream.read())
 
 
 def patch_and_save_config(config, patch):
